--- a/security/pkg/k8s/chiron/controller.go
+++ b/security/pkg/k8s/chiron/controller.go
@@ -62,12 +62,15 @@ const (
 	// The number of retries when requesting to create secret.
 	secretCreationRetry = 3
 
-	// The interval for reading a certificate
-	certReadInterval = 500 * time.Millisecond
 	// The number of tries for reading a certificate
 	maxNumCertRead = 10
-	// timeout for reading signed CSR
-	timeoutForReadingCSR = 5 * time.Second
+
+	// The interval for reading a certificate
+	certReadInterval = 500 * time.Millisecond
+)
+
+var (
+	certWatchTimeout = 5 * time.Second
 )
 
 // WebhookController manages the service accounts' secrets that contains Istio keys and certificates.