--- a/security/pkg/k8s/chiron/controller_test.go
+++ b/security/pkg/k8s/chiron/controller_test.go
@@ -189,7 +189,7 @@ func TestUpsertSecret(t *testing.T) {
 			},
 		}
 		client.PrependReactor("get", "certificatesigningrequests", defaultReactionFunc(csr))
-
+		certWatchTimeout = time.Millisecond
 		wc, err := NewWebhookController(tc.gracePeriodRatio, tc.minGracePeriod,
 			client.CoreV1(), client.AdmissionregistrationV1beta1(), client.CertificatesV1beta1(),
 			tc.k8sCaCertFile, tc.secretNames, tc.dnsNames, tc.serviceNamespaces)
@@ -239,6 +239,7 @@ func TestScrtDeleted(t *testing.T) {
 		},
 	}
 
+	certWatchTimeout = time.Millisecond
 	for _, tc := range testCases {
 		client := fake.NewSimpleClientset()
 		client.PrependReactor("get", "certificatesigningrequests", defaultReactionFunc(csr))
@@ -288,6 +289,7 @@ func TestScrtDeleted(t *testing.T) {
 func TestScrtUpdated(t *testing.T) {
 	dnsNames := []string{"foo"}
 
+	certWatchTimeout = time.Millisecond
 	testCases := map[string]struct {
 		gracePeriodRatio       float32
 		minGracePeriod         time.Duration