Before:
istio.io/istio/security/pkg/pki/ca      96.652s
istio.io/istio/security/pkg/k8s/chiron  74.827s

After:
istio.io/istio/security/pkg/pki/ca      4.063s
istio.io/istio/security/pkg/k8s/chiron  1.602s



[ ] Configuration Infrastructure
[ ] Docs
[ ] Installation
[ ] Networking
[ ] Performance and Scalability
[ ] Policies and Telemetry
[ ] Security
[ ] Test and Release
[ ] User Experience
[ ] Developer Infrastructure